+++ Java/src/main/java/SiiBoleta/RNG.java
@@ -4,12 +4,12 @@ import java.math.BigInteger;
 
 public class RNG {
 
-    public static DTEDefType.Documento.TED.DD.CAF.DA.RNG makeRNG(BigInteger d,
-                                                   BigInteger h) {
+    public static DTEDefType.Documento.TED.DD.CAF.DA.RNG makeRNG(int d,
+                                                   int h) {
 
         DTEDefType.Documento.TED.DD.CAF.DA.RNG rng = new DTEDefType.Documento.TED.DD.CAF.DA.RNG();
-        rng.setD(d);
-        rng.setH(h);
+        rng.setD(BigInteger.valueOf(d));
+        rng.setH(BigInteger.valueOf(h));
         return rng;
     }
 }
